// src/app/models/activity.model.ts
export class Activity {
  id?: any;
  title?:string;
  description?: string;
  address?: string;
  ville?: string;
  codePostal?: string;
  governorate?: string;
  codeFiscal?: string;
  photo?: string;
  email?: string;
  numTel?: string;
  siteWeb?: string;
  workerNumber?: number;
  amount?: number;
  beneficiaryCin?: string;
  dossierTitle?: string;
  published?: boolean;
  creationDate?:string;
}